+++ widgetbook/lib/simple_draggable_builder.dart
@@ -1,26 +1,27 @@
 import 'package:draggable_builder/draggable_builder.dart';
 import 'package:flutter/material.dart';
 import 'package:widgetbook_annotation/widgetbook_annotation.dart' as widgetbook;
+import 'package:widgetbook_workspace/widgets/grid_view.dart';
 
 @widgetbook.UseCase(
   name: 'Simple',
   type: DraggableBuilder,
 )
-Widget buildDraggableBuilderUseCase(BuildContext context) {
-  return DraggableBuilderUseCase();
+Widget buildSimpleDraggableBuilderUseCase(BuildContext context) {
+  return SimpleDraggableBuilderUseCase();
 }
 
-class DraggableBuilderUseCase extends StatefulWidget {
-  const DraggableBuilderUseCase({super.key});
+class SimpleDraggableBuilderUseCase extends StatefulWidget {
+  const SimpleDraggableBuilderUseCase({super.key});
 
   @override
-  State<DraggableBuilderUseCase> createState() => _DraggableBuilderUseCaseState();
+  State<SimpleDraggableBuilderUseCase> createState() => _SimpleDraggableBuilderUseCaseState();
 }
 
-class _DraggableBuilderUseCaseState extends State<DraggableBuilderUseCase> {
-  var _colors = [Colors.red, Colors.yellow, Colors.green, Colors.blue];
-
+class _SimpleDraggableBuilderUseCaseState extends State<SimpleDraggableBuilderUseCase> {
   late final DraggableController _controller;
+
+  var _colors = [Colors.red, Colors.yellow, Colors.green, Colors.blue];
 
   @override
   void initState() {
@@ -39,27 +40,28 @@ class _DraggableBuilderUseCaseState extends State<DraggableBuilderUseCase> {
 
   @override
   Widget build(BuildContext context) {
-    return DraggableBuilder(
-      controller: _controller,
-      itemBuilder: (_, index) => ColoredBox(color: _colors[index]),
-      itemCount: _colors.length,
-      builder: (_, itemBuilder, itemCount) => GridView.builder(
-        shrinkWrap: true,
-        gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-          crossAxisCount: 4,
-          mainAxisSpacing: 8.0,
-          crossAxisSpacing: 8.0,
+    return SingleChildScrollView(
+      child: Padding(
+        padding: EdgeInsets.all(8),
+        child: DraggableBuilder(
+          controller: _controller,
+          itemBuilder: (_, index) => ColoredBox(color: _colors[index]),
+          itemCount: _colors.length,
+          builder: (_, itemBuilder, itemCount) => MyGridView(
+            itemBuilder: itemBuilder,
+            itemCount: itemCount,
+          ),
         ),
-        itemBuilder: itemBuilder,
-        itemCount: itemCount,
       ),
     );
   }
 
   void _onDragCompletion(Object dragId, int dragIndex, Object targetId, int targetIndex) {
+    final color = _colors[dragIndex];
+
     final newColors = [..._colors] //
       ..removeAt(dragIndex)
-      ..insert(targetIndex, _colors[dragIndex]);
+      ..insert(targetIndex, color);
 
     setState(() {
       _colors = newColors;