+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.0.2
 
+* Mention support for multiple draggable containers and item exchange
+
 ## 0.0.1
 
 * Add DraggableBuilder widget